+++ src/DependencyInjection/Configuration.php
@@ -134,6 +134,52 @@ class Configuration implements ConfigurationInterface
 
                     throw new InvalidConfigurationException('To enable the user context logout handler, you need to configure a ban capable proxy_client.');
                 })
+            ->end()
+            // Determine the default tags header for the varnish client, depending on whether we use BAN or xkey
+            ->validate()
+                ->ifTrue(
+                    function ($v) {
+                        return
+                            array_key_exists('proxy_client', $v)
+                            && array_key_exists('varnish', $v['proxy_client'])
+                            && empty($v['proxy_client']['varnish']['tags_header'])
+                        ;
+                    }
+                )
+                ->then(function ($v) {
+                    $v['proxy_client']['varnish']['tags_header'] =
+                        (Varnish::TAG_XKEY === $v['proxy_client']['varnish']['tag_mode'])
+                        ? Varnish::DEFAULT_HTTP_HEADER_CACHE_XKEY
+                        : Varnish::DEFAULT_HTTP_HEADER_CACHE_TAGS;
+
+                    return $v;
+                })
+            ->end()
+            // Determine the default tag response header, depending on whether we use BAN or xkey
+            ->validate()
+                ->ifTrue(
+                    function ($v) {
+                        return empty($v['tags']['response_header']);
+                    }
+                )
+                ->then(function ($v) {
+                    $v['tags']['response_header'] = $this->isVarnishXkey($v) ? 'xkey' : TagHeaderFormatter::DEFAULT_HEADER_NAME;
+
+                    return $v;
+                })
+            ->end()
+            // Determine the default separator for the tags header, depending on whether we use BAN or xkey
+            ->validate()
+                ->ifTrue(
+                    function ($v) {
+                        return empty($v['tags']['separator']);
+                    }
+                )
+                ->then(function ($v) {
+                    $v['tags']['separator'] = $this->isVarnishXkey($v) ? ' ' : ',';
+
+                    return $v;
+                })
         ;
 
         $this->addCacheableResponseSection($rootNode);
@@ -148,6 +194,14 @@ class Configuration implements ConfigurationInterface
         $this->addDebugSection($rootNode);
 
         return $treeBuilder;
+    }
+
+    private function isVarnishXkey(array $v): bool
+    {
+        return array_key_exists('proxy_client', $v)
+            && array_key_exists('varnish', $v['proxy_client'])
+            && Varnish::TAG_XKEY === $v['proxy_client']['varnish']['tag_mode']
+        ;
     }
 
     private function addCacheableResponseSection(ArrayNodeDefinition $rootNode)
@@ -362,7 +416,6 @@ class Configuration implements ConfigurationInterface
                             ->end()
                             ->children()
                                 ->scalarNode('tags_header')
-                                    ->defaultValue(Varnish::DEFAULT_HTTP_HEADER_CACHE_TAGS)
                                     ->info('HTTP header to use when sending tag invalidation requests to Varnish')
                                 ->end()
                                 ->scalarNode('header_length')
@@ -372,6 +425,11 @@ class Configuration implements ConfigurationInterface
                                     ->useAttributeAsKey('name')
                                     ->info('Map of additional headers to include in each ban request.')
                                     ->prototype('scalar')->end()
+                                ->end()
+                                ->enumNode('tag_mode')
+                                    ->info('If you can enable the xkey module in Varnish, use the purgekeys mode for more efficient tag handling')
+                                    ->values(['ban', 'purgekeys'])
+                                    ->defaultValue('ban')
                                 ->end()
                                 ->append($this->getHttpDispatcherNode())
                             ->end()
@@ -580,8 +638,12 @@ class Configuration implements ConfigurationInterface
                             ->info('Service name of a custom ExpressionLanugage to use.')
                         ->end()
                         ->scalarNode('response_header')
-                            ->defaultValue(TagHeaderFormatter::DEFAULT_HEADER_NAME)
-                            ->info('HTTP header that contains cache tags')
+                            ->defaultNull()
+                            ->info('HTTP header that contains cache tags. Defaults to xkey-softpurge for Varnish xkey or X-Cache-Tags otherwise')
+                        ->end()
+                        ->scalarNode('separator')
+                            ->defaultNull()
+                            ->info('Character(s) to use to separate multiple tags. Defaults to " " for Varnish xkey or "," otherwise')
                         ->end()
                         ->arrayNode('rules')
                             ->prototype('array')
@@ -593,8 +655,8 @@ class Configuration implements ConfigurationInterface
                                     })
                                     ->thenInvalid('Configured a tag_expression but ExpressionLanugage is not available')
                                 ->end()
-                                ->children();
-
+                                ->children()
+                        ;
         $this->addMatch($rules);
 
         $rules